+++ .github/workflows/test.yml
@@ -13,8 +13,6 @@ jobs:
         run: curl -sSfL https://raw.githubusercontent.com/golangci/golangci-lint/master/install.sh | sh -s -- -b $(go env GOPATH)/bin v1.30.0
       - name: lint
         run: $(go env GOPATH)/bin/golangci-lint run ./...
-      - name: test
-        run: go test ./... -v
       # test with firestore
       - name: docker-compose up
         run: docker-compose up -d
@@ -26,6 +24,4 @@ jobs:
           FIRESTORE_EMULATOR_HOST: 0.0.0.0:8080
       - name: go test with firestore
         run: |
-          go test ./... -v -tags=integration
-        env:
-          FIRESTORE_EMULATOR_HOST: 0.0.0.0:8080
+          FIRESTORE_EMULATOR_HOST=0.0.0.0:8080 go test ./... -v
