+++ testData/typelist.json
@@ -1,0 +1,1 @@
+{"geo":{"latitude":80,"longitude":80},"nil":null,"ref":{"Parent":{"Parent":null,"Path":"projects/fsrpldemo/databases/(default)/documents/dest","ID":"dest"},"Path":"projects/fsrpldemo/databases/(default)/documents/dest/dog","ID":"dog"},"refSub":{"Parent":{"Parent":{"Parent":{"Parent":null,"Path":"projects/fsrpldemo/databases/(default)/documents/original","ID":"original"},"Path":"projects/fsrpldemo/databases/(default)/documents/original/dog","ID":"dog"},"Path":"projects/fsrpldemo/databases/(default)/documents/original/dog/subCol","ID":"subCol"},"Path":"projects/fsrpldemo/databases/(default)/documents/original/dog/subCol/subDoc","ID":"subDoc"},"str":"string","time":"2020-08-31T15:00:00Z"}
