+++ testData/user.json
@@ -1,0 +1,1 @@
+{"_created_at":"2019-08-26T05:00:00Z","coin":0,"favorites":["1","2"],"isDeleted":true,"mapData":{"b":true,"name":"mName"},"name":"user"}
